cdk/requirements.txt
aws-cdk-lib>=2.120.0
constructs>=10.0.0
boto3>=1.34.0
aws-cdk.aws-lambda-python-alpha>=2.120.0-alpha.0
typing-extensions>=4.9.0